SD card Files Handeling and init works
