+++ SPI SD Card.lnp
@@ -4,6 +4,7 @@
 ".\main.o"
 ".\spi.o"
 ".\uart.o"
+".\sd_card.o"
 --ro-base 0x00000000 --entry 0x00000000 --rw-base 0x20000000 --entry Reset_Handler --first __Vectors --strict --summary_stderr --info summarysizes --map --xref --callgraph --symbols
 --info sizes --info totals --info unused --info veneers
 --list ".\SPI SD Card.map" -o ".\SPI SD Card.axf"
