+++ main.d
@@ -2,6 +2,8 @@
 .\main.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\inttypes.h
 .\main.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdint.h
 .\main.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdio.h
+.\main.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\time.h
 .\main.o: src\../include/spi.h
 .\main.o: src\../include/gpio.h
 .\main.o: src\../include/uart.h
+.\main.o: src\../include/sd_card.h
