+++ sd_card.d
@@ -1,0 +1,9 @@
+.\sd_card.o: src\sd_card.c
+.\sd_card.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdio.h
+.\sd_card.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\inttypes.h
+.\sd_card.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdint.h
+.\sd_card.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\time.h
+.\sd_card.o: src\../include/spi.h
+.\sd_card.o: src\../include/gpio.h
+.\sd_card.o: src\../include/uart.h
+.\sd_card.o: src\../include/sd_card.h
