+++ spi.d
@@ -1,5 +1,6 @@
 .\spi.o: src\spi.c
 .\spi.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\inttypes.h
 .\spi.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdint.h
+.\spi.o: D:\Keil uVision4\ARM\ARMCC\bin\..\include\stdio.h
 .\spi.o: src\../include/spi.h
 .\spi.o: src\../include/uart.h
